add door lock command and lock state attribute

+++ zigbee-api/src/main/java/org/bubblecloud/zigbee/api/cluster/impl/DoorLockImpl.java
@@ -100,7 +100,7 @@ public class DoorLockImpl implements DoorLock {
 	
 	public int getLockState() throws ZigBeeDeviceException {
 		 try {
-	            return (int) doorLockCluster.getAttributeLockState().getValue();
+	            return (int) (Integer) doorLockCluster.getAttributeLockState().getValue();
 	        } catch (ZigBeeClusterException e) {
 	            throw new ZigBeeDeviceException(e);
 	        }
